Move `eval_body_using_ecx` to the only module it is used in
